Moved tech component into slices. Added case studies header and tech slice
