+++ site/slices/tech-slice/index.js
