+++ site/slices/tech-slice/styles.css
@@ -1,5 +1,5 @@
-@value black, white from "../../../css/_colors.css";
-@value maxWidth, largeScreen, mediumScreen from "../../../css/_breakpoints.css";
+@value black, white from "../../css/_colors.css";
+@value maxWidth, largeScreen, mediumScreen from "../../css/_breakpoints.css";
 
 .widthContainer {
   max-width: maxWidth;
@@ -8,7 +8,7 @@
 }
 
 .techSlice {
-  composes: dividerBlack from "../../../css/_divider.css";
+  composes: dividerBlack from "../../css/_divider.css";
   color: white;
   background-color: black;
   padding: 30px 20px 40px 20px;
@@ -17,17 +17,16 @@
 
 .heading {
   composes: widthContainer;
-  composes: serif from "../../../css/typography/_fonts.css";
-  composes: fontL from "../../../css/typography/_fonts.css";
+  composes: serif from "../../css/typography/_fonts.css";
+  composes: fontL from "../../css/typography/_fonts.css";  text-align: center;
   font-weight: bold;
-  text-align: center;
   margin-bottom: 20px;
 }
 
 .projectsBlerb {
   composes: widthContainer;
-  composes: serif from "../../../css/typography/_fonts.css";
-  composes: fontS from "../../../css/typography/_fonts.css";
+  composes: serif from "../../css/typography/_fonts.css";
+  composes: fontS from "../../css/typography/_fonts.css";
   font-weight: bold;
   text-align: center;
   margin-bottom: 50px;
@@ -57,12 +56,12 @@
 }
 
 .techName {
-  composes: boldSansSerif from "../../../css/typography/_fonts.css";
+  composes: boldSansSerif from "../../css/typography/_fonts.css";
   margin: 20px 0px 40px 0px;
 }
 
 .moreBtn {
-  composes: primaryButton from "../../../css/_links.css";
+  composes: primaryButton from "../../css/_links.css";
 }
 
 @media mediumScreen {
@@ -75,8 +74,8 @@
   }
 
   .projectsBlerb {
-    composes: fontSerif from "../../../css/typography/_fonts.css";
-    composes: fontS from "../../../css/typography/_fonts.css";
+    composes: fontSerif from "../../css/typography/_fonts.css";
+    composes: fontS from "../../css/typography/_fonts.css";
     margin-bottom: 90px;
   }
 
@@ -91,6 +90,7 @@
   .moreBtn {
     composes: primaryButton from "../../../css/_links.css";
     margin: 60px 0 0 0;
+    composes: primaryButton from "../../css/_links.css";
   }
 }
